pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seqthreat"
version = "0.1.0"
description = "Sequence-based threat detection using character n-gram feature engineering"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Patrick Grady"}]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "scipy>=1.11",
    "click>=8.1",
    "pyyaml>=6.0",
    "mlflow>=2.10",
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "joblib>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "httpx>=0.27",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
]

[project.scripts]
seqthreat = "src.cli:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
